caching container client instead of sas token
